fixed rating on detail page

+++ src/app/shared/button/button.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { Router } from '@angular/router';
+import { sample } from 'rxjs';
 import { Liked } from 'src/app/core/models/liked';
 import { Movie } from 'src/app/core/models/movie';
 import { LikedService } from 'src/app/core/services/liked.service';
@@ -13,7 +14,7 @@ import { MovieService } from '../movie/movie.service';
 export class ButtonComponent implements OnInit {
   @Input() movie!: Movie;
   @Input() liked!: Liked[];
-  saved: boolean = false;
+  isSaved: boolean = false;
 
   constructor(
     private movieService: MovieService,
@@ -22,30 +23,35 @@ export class ButtonComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    this.checkSaved();
+    // need this timeout because the @Input has a slight delay over init
+    setTimeout(() => {
+      this.checkSaved();
+    }, 40);
   }
 
   checkSaved(): void {
-    if (this.liked.filter((l) => l.id == this.movie.id).length == 0) {
-      this.saved = true;
+    // check that the current movie is in the liked movies array.
+    let ids: number[] = [];
+
+    this.liked.forEach((l) => {
+      ids.push(l.id);
+    });
+    // if so: set the saved to true (because the movie is found in the array).
+    if (ids.includes(this.movie.id)) {
+      this.isSaved = true;
     }
   }
 
-  route(event: any) {
+  toggleSaved(event: any): void {
     event.stopPropagation();
-
-    // this.router.navigate(["/searchlist"]);
-  }
-
-  save(event: any): void {
-    event.stopPropagation();
-    this.likedService.postLikedMovie(this.movie, 0, '');
-    console.log('saved', this.movie);
-  }
-
-  remove(event: any): void {
-    event.stopPropagation();
-    this.likedService.deleteLikedMovie(this.movie.id);
-    console.log('removed', this.movie);
+    if (this.isSaved) {
+      this.likedService.deleteLikedMovie(this.movie.id);
+      console.log('removed', this.movie);
+    } else {
+      this.likedService.postLikedMovie(this.movie, 0, '');
+      console.log('saved', this.movie);
+    }
+    // set the state for event binding, it will only check on initialize.
+    this.isSaved = !this.isSaved;
   }
 }
